불필요 컴포넌트 정리
